Restructure and internalize (#1738)
* Move Reference to Jint.Runtime namespace
* Internalize Binding
* Internalize Advanced operations constructors
* Make Advanced operations inner classes
* Make Options sub-options inner classes
* Move delegate definitions under options
* Make StrictModeScope internal
* Move ModuleBuilder under Runtime.Modules namespace
