+++ Jint.Tests/Runtime/NullPropagation.cs
@@ -1,7 +1,6 @@
 ﻿using Jint.Native;
 using Jint.Runtime;
 using Jint.Runtime.Interop;
-using Jint.Runtime.References;
 
 namespace Jint.Tests.Runtime
 {
