+++ Jint/Runtime/Interpreter/Expressions/BindingPatternAssignmentExpression.cs
@@ -3,7 +3,6 @@ using Jint.Native;
 using Jint.Native.Array;
 using Jint.Native.Function;
 using Jint.Native.Iterator;
-using Jint.Runtime.References;
 using Environment = Jint.Runtime.Environments.Environment;
 
 namespace Jint.Runtime.Interpreter.Expressions
