+++ Jint/Runtime/Interpreter/Expressions/JintTaggedTemplateExpression.cs
@@ -2,7 +2,6 @@ using Esprima.Ast;
 using Jint.Native;
 using Jint.Native.Object;
 using Jint.Runtime.Descriptors;
-using Jint.Runtime.References;
 
 namespace Jint.Runtime.Interpreter.Expressions;
